6. practice

+++ 2021-22-2/webeng5/mytracks/app/Http/Controllers/ProjectController.php
@@ -5,17 +5,19 @@ namespace App\Http\Controllers;
 use App\Http\Requests\ProjectFormRequest;
 use App\Models\Project;
 use Illuminate\Http\Request;
+use Auth;
 
 class ProjectController extends Controller
 {
     public function index() {
-        $projects = Project::all();
+        // $projects = Project::all();
         return view('projects.list', [
-            'projects' => $projects,
+            'projects' => Auth::user()->projects,
         ]); // projects/list.blade.php
     }
 
     public function show(Project $project) {
+        $this->authorize('access', $project);
         return view('projects.show', [
             'project'   => $project,
         ]); // projects/show.blade.php
@@ -26,22 +28,25 @@ class ProjectController extends Controller
     }
 
     public function store(ProjectFormRequest $request) {
-        Project::create($request->validated());
+        Auth::user()->projects()->create($request->validated());
         return redirect()->route('projects.index'); // GET
     }
 
     public function edit(Project $project) {
+        $this->authorize('access', $project);
         return view('projects.edit', [
             'project'   => $project,
         ]); // projects/edit.blade.php
     }
 
     public function update(Project $project, ProjectFormRequest $request) {
+        $this->authorize('access', $project);
         $project->update($request->validated());
         return redirect('/projects'); // GET
     }
 
     public function destroy(Project $project) {
+        $this->authorize('access', $project);
         $project->delete();
         return redirect('/projects');
     }
